requirements.txt
GitPython==3.1.44
Jinja2==3.1.6
MarkupSafe==3.0.2
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.11.2
cycler==0.12.1
fonttools==4.58.2
frozendict==2.4.6
gitdb==4.0.12
holidays==0.74
idna==3.10
importlib_resources==6.5.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
narwhals==1.41.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prophet==1.1.7
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
smmap==6.0.0
soupsieve==2.7
stanio==0.5.1
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.46.1
yfinance==0.2.61
